[Feature] Add the definition of lynx_value and related interfaces
Add the definition of lynx_value and related interfaces to accessing lynx_value. The lynx_value is a generic Value container that will act as a bridge for cross-module communication. The interface to access lynx_value is placed in lynx_api_env. The interfaces in lynx_api_env can have different implementations. The implementation will be submitted in a subsequent PR.

issue:m-6042769122