[Optimize] __UpdateComponentInfo supports updating the Config of ComponentElement.
As mentioned in the title, let __UpdateComponentInfo support updating the Config of ComponentElement.
When updating the ComponentElement Config, check if config.removeComponentElement is true. If it is true, then call MarkAsWrapperComponent.

At the same time, the logic of MarkAsWrapperComponent has been modified to ensure that MarkAsWrapperComponent can only be called before creating the ComponentElement and resolving it. It does not support dynamic switching between Wrapper and non-Wrapper ComponentElements.

issue:m-6083199326
AutoSubmit:True
AutoLand:release/3.2