[BugFix][Event] Fix the problem of triggering tap event when the foldview is scrolling by inertia.
description: When you click on the foldview to stop the inertial scrolling, the tap event should not be triggered. Therefore, on the iOS side, you need to set lynxEnableTapGestureSimultaneously to false during scrolling, and on the Android side, you cannot clear the related UISet when the finger is pressed.

AutoLand: release/3.2

issue: m-5736536125