[Optimize][Element] Integrate Layout in Element for Embedded Mode
In this merge request, for Embedded Mode, a LayoutObject is added to FiberElement to minimize the overhead of copying computed_css_style and building the LayoutNode tree. Consequently, the layout process, which includes updating layout information and initiating LayoutNode patching, is handled within the PageElement layout procedure.

Additionally, a new mode is introduced in Embedded Mode to manage the activation of Layout-Element integration.

AutoSubmit: true

issue:m-0