[Optimize][LynxRecorder][Reland] Remove unnecessary recording points
The current changes optimize LynxRecorder for script recording in Scripts, including the following two points:

1. Removing redundant recording points. We can now fully cover all script loading processes in runtime_mediator
without the need for additional operations in js_app.

2. Changing recordScript from a broadcast behavior to unicast behavior. Previously, without injecting record_id
in runtime_mediator, we couldn't differentiate pages, so we had to record broadcasts and store for each page.
Now, we inject record_id to indicate which page this operation is from.

SkipChecks:macro