[Optimize] Make some lepus benchmark cases more stable
Case removed. The codes that actually work in the performance test of these two use cases are very light in logic, and the test case data results are unstable, so they can be deleted.

  BM_ToLepusValueShallowConvert
  BM_FromLepusShallowConvert

Cases modified. Optimized some test cases to exclude the destruction time of local objects from the test time.

  BM_ToLepusValueDeepConvert
  BM_TestEmptyRenderNGFunction