[Optimize] Implement default constructor for Vector and sub-types
Previously, the default constructors for Vector and its subclasses were set to be implemented by the compiler(by `= default`). In this case, when using the "new T()" constructor, the compiler would default to performing value-initialization and setting sizeof(T) to all zeros with instructions or `bzero` or `memset` external calls. This behavior was unexpected and could lead to performance degradation. This was especially true for InlineType<N>, which holds an internal buffer, potentially resulting in a larger sizeof(T).